requirements.txt
apache-airflow==2.6.3
undetected-chromedriver>=3.5.0
selenium>=4.10.0
pyvirtualdisplay>=3.0
webdriver-manager>=3.8.6 
pandas>=1.5.3
numpy>=1.23.5
mysqlclient>=2.1.1